refactor listings offer to use GenericOffer
